Reorganizando as pastas da seção de informações
